Fixed: remove some test that use the selector ancestor::
The <blockquote> in list items forgets to translate its internal <para> because
there is a non-related ancestor of type list that prevent the add of <t> tags.

Instead: prevent wrapping the <blockquote> with a <t> if the parent item
is a listitem.

Note: listitem suits better since it is used for orderedlist, itemizedlist and variablelist